Added snapshot tests for MBWay component
