+++ Tests/Components Tests/MB Way/MBWayComponentTests.swift
@@ -61,98 +61,6 @@ class MBWayComponentTests: XCTestCase {
         XCTAssertEqual(sut.button.title, localizedString(LocalizationKey(key: "adyen_continueTo"), sut.configuration.localizationParameters, paymentMethod.name))
     }
 
-    func testUIConfiguration() {
-        var style = FormComponentStyle()
-
-        /// Footer
-        style.mainButtonItem.button.title.color = .white
-        style.mainButtonItem.button.title.backgroundColor = .red
-        style.mainButtonItem.button.title.textAlignment = .center
-        style.mainButtonItem.button.title.font = .systemFont(ofSize: 22)
-        style.mainButtonItem.button.backgroundColor = .red
-        style.mainButtonItem.backgroundColor = .brown
-
-        /// background color
-        style.backgroundColor = .red
-
-        /// Text field
-        style.textField.text.color = .red
-        style.textField.text.font = .systemFont(ofSize: 13)
-        style.textField.text.textAlignment = .right
-
-        style.textField.title.backgroundColor = .blue
-        style.textField.title.color = .yellow
-        style.textField.title.font = .systemFont(ofSize: 20)
-        style.textField.title.textAlignment = .center
-        style.textField.backgroundColor = .red
-
-        let config = MBWayComponent.Configuration(style: style)
-        let sut = MBWayComponent(paymentMethod: paymentMethod,
-                                 context: context,
-                                 configuration: config)
-
-        UIApplication.shared.keyWindow?.rootViewController = sut.viewController
-
-        wait(for: .milliseconds(300))
-        
-        let phoneNumberView: FormPhoneNumberItemView? = sut.viewController.view.findView(with: MBWayViewIdentifier.phone)
-        let phoneNumberViewTitleLabel: UILabel? = sut.viewController.view.findView(with: MBWayViewIdentifier.phoneTitleLabel)
-        let phoneNumberViewTextField: UITextField? = sut.viewController.view.findView(with: MBWayViewIdentifier.phoneTextField)
-
-        let payButtonItemViewButton: UIControl? = sut.viewController.view.findView(with: MBWayViewIdentifier.payButton)
-        let payButtonItemViewButtonTitle: UILabel? = sut.viewController.view.findView(with: MBWayViewIdentifier.payButtonTitleLabel)
-
-        /// Test phone number field
-        XCTAssertEqual(phoneNumberView?.backgroundColor, .red)
-        XCTAssertEqual(phoneNumberViewTitleLabel?.textColor, sut.viewController.view.tintColor)
-        XCTAssertEqual(phoneNumberViewTitleLabel?.backgroundColor, .blue)
-        XCTAssertEqual(phoneNumberViewTitleLabel?.textAlignment, .center)
-        XCTAssertEqual(phoneNumberViewTitleLabel?.font, .systemFont(ofSize: 20))
-        XCTAssertEqual(phoneNumberViewTextField?.backgroundColor, .red)
-        XCTAssertEqual(phoneNumberViewTextField?.textAlignment, .right)
-        XCTAssertEqual(phoneNumberViewTextField?.textColor, .red)
-        XCTAssertEqual(phoneNumberViewTextField?.font, .systemFont(ofSize: 13))
-
-        /// Test footer
-        XCTAssertEqual(payButtonItemViewButton?.backgroundColor, .red)
-        XCTAssertEqual(payButtonItemViewButtonTitle?.backgroundColor, .red)
-        XCTAssertEqual(payButtonItemViewButtonTitle?.textAlignment, .center)
-        XCTAssertEqual(payButtonItemViewButtonTitle?.textColor, .white)
-        XCTAssertEqual(payButtonItemViewButtonTitle?.font, .systemFont(ofSize: 22))
-    }
-
-    func testSubmitForm() throws {
-        let sut = MBWayComponent(paymentMethod: paymentMethod, context: context)
-        let delegate = PaymentComponentDelegateMock()
-        sut.delegate = delegate
-
-        let delegateExpectation = expectation(description: "PaymentComponentDelegate must be called when submit button is clicked.")
-        delegate.onDidSubmit = { data, component in
-            XCTAssertTrue(component === sut)
-            XCTAssertTrue(data.paymentMethod is MBWayDetails)
-            let data = data.paymentMethod as! MBWayDetails
-            XCTAssertEqual(data.telephoneNumber, "+3511233456789")
-
-            sut.stopLoadingIfNeeded()
-            delegateExpectation.fulfill()
-            XCTAssertEqual(sut.viewController.view.isUserInteractionEnabled, true)
-            XCTAssertEqual(sut.button.showsActivityIndicator, false)
-        }
-
-        UIApplication.shared.keyWindow?.rootViewController = sut.viewController
-        
-        wait(for: .milliseconds(300))
-        
-        let submitButton: UIControl? = sut.viewController.view.findView(with: MBWayViewIdentifier.payButton)
-
-        let phoneNumberView: FormPhoneNumberItemView! = sut.viewController.view.findView(with: MBWayViewIdentifier.phone)
-        self.populate(textItemView: phoneNumberView, with: "1233456789")
-
-        submitButton?.sendActions(for: .touchUpInside)
-
-        waitForExpectations(timeout: 10, handler: nil)
-    }
-
     func testBigTitle() {
         let sut = MBWayComponent(paymentMethod: paymentMethod,
                                  context: context)
